refactor(core): move `ViewEncapsulation` and `ViewType` to the right places
