+++ modules/angular2/src/core/metadata/view.ts
@@ -1,7 +1,36 @@
 import {CONST, Type} from 'angular2/src/core/facade/lang';
-import {ViewEncapsulation} from 'angular2/src/core/render/api';
 
-export {ViewEncapsulation} from 'angular2/src/core/render/api';
+/**
+ * Defines template and style encapsulation options available for Component's {@link View}.
+ *
+ * See {@link ViewMetadata#encapsulation}.
+ */
+export enum ViewEncapsulation {
+  /**
+   * Emulate `Native` scoping of styles by adding an attribute containing surrogate id to the Host
+   * Element and pre-processing the style rules provided via
+   * {@link ViewMetadata#styles} or {@link ViewMetadata#stylesUrls}, and adding the new Host Element
+   * attribute to all selectors.
+   *
+   * This is the default option.
+   */
+  Emulated,
+  /**
+   * Use the native encapsulation mechanism of the renderer.
+   *
+   * For the DOM this means using [Shadow DOM](https://w3c.github.io/webcomponents/spec/shadow/) and
+   * creating a ShadowRoot for Component's Host Element.
+   */
+  Native,
+  /**
+   * Don't provide any template or style encapsulation.
+   */
+  None
+}
+
+export var VIEW_ENCAPSULATION_VALUES =
+    [ViewEncapsulation.Emulated, ViewEncapsulation.Native, ViewEncapsulation.None];
+
 
 /**
  * Metadata properties available for configuring Views.
