+++ modules_dart/transform/lib/src/transform/common/directive_metadata_reader.dart
@@ -7,7 +7,7 @@ import 'package:angular2/src/core/compiler/directive_metadata.dart';
 import 'package:angular2/src/core/compiler/template_compiler.dart';
 import 'package:angular2/src/core/change_detection/change_detection.dart';
 import 'package:angular2/src/core/linker/interfaces.dart' show LifecycleHooks;
-import 'package:angular2/src/core/render/api.dart' show ViewEncapsulation;
+import 'package:angular2/src/core/metadata/view.dart' show ViewEncapsulation;
 import 'package:angular2/src/transform/common/annotation_matcher.dart';
 import 'package:angular2/src/transform/common/interface_matcher.dart';
 import 'package:barback/barback.dart' show AssetId;
